test: split test_libmrot into multiple test targets
